requirements.txt
asgiref==3.6.0
Django==3.2.16
django-bootstrap-modal-forms==2.2.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
Pillow==9.4.0
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.4.0
